requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.0.1 